LayoutTests/platform/ios/imported/w3c/web-platform-tests/editing/other/restyle-textcontrol-during-beforeinput-expected.txt



FAIL input should accept text input even after restyled assert_equals: expected "ab" but got ""
FAIL textarea should accept text input even after restyled assert_equals: expected "ab" but got ""
